API Reference

Exports

GeometryBasics.AbstractGeometry — Type

Abstract Geometry in R{Dim} with Number type T

GeometryBasics.AbstractMesh — Type

AbstractMesh

An abstract mesh is a collection of Polytope elements (Simplices / Ngons). The connections are defined via faces(mesh), the coordinates of the elements are returned by coordinates(mesh). Arbitrary meta information can be attached per point or per face

GeometryBasics.Circle — Type

Circle{T}

An alias for a HyperSphere of dimension 2. (i.e. HyperSphere{2, T})

GeometryBasics.Cylinder — Type

Cylinder{N, T}

A Cylinder is a 2D rectangle or a 3D cylinder defined by its origin point, its extremity and a radius. origin, extremity and r, must be specified.

GeometryBasics.Cylinder2 — Type

Cylinder2{T}
Cylinder3{T}

A Cylinder2 or Cylinder3 is a 2D/3D cylinder defined by its origin point, its extremity and a radius. origin, extremity and r, must be specified.

GeometryBasics.HyperRectangle — Type

HyperRectangle{N, T}

A HyperRectangle is a generalization of a rectangle into N-dimensions. Formally it is the cartesian product of intervals, which is represented by the origin and widths fields, whose indices correspond to each of the N axes.

GeometryBasics.HyperSphere — Type

HyperSphere{N, T}

A HyperSphere is a generalization of a sphere into N-dimensions. A center and radius, r, must be specified.

GeometryBasics.LineString — Type

LineString(points::AbstractVector{<:Point})

A LineString is a geometry of connected line segments

GeometryBasics.Mesh — Type

Mesh <: AbstractMesh{Element}

The concrete AbstractMesh type.

GeometryBasics.MultiPoint — Type

MultiPoint(points::AbstractVector{AbstractPoint})

A collection of points

GeometryBasics.MultiPolygon — Type

MultiPolygon(polygons::AbstractPolygon)

GeometryBasics.OffsetInteger — Type

OffsetInteger{O, T}

OffsetInteger type mainly for indexing.

O - The offset relative to Julia arrays. This helps reduce copying when

communicating with 0-indexed systems such as OpenGL.

GeometryBasics.Polygon — Type

Polygon(exterior::AbstractVector{<:Point})
Polygon(exterior::AbstractVector{<:Point}, interiors::Vector{<:AbstractVector{<:Point}})

GeometryBasics.Polytope — Type

Geometry made of N connected points. Connected as one flat geometry, it makes a Ngon / Polygon. Connected as volume it will be a Simplex / Tri / Cube. Note That Polytope{N} where N == 3 denotes a Triangle both as a Simplex or Ngon.

GeometryBasics.Polytope — Method

The Ngon Polytope element type when indexing an array of points with a SimplexFace

GeometryBasics.Polytope — Method

The Simplex Polytope element type when indexing an array of points with a SimplexFace

GeometryBasics.Polytope — Method

The fully concrete Ngon type, when constructed from a point type!

GeometryBasics.Polytope — Method

The fully concrete Simplex type, when constructed from a point type!

GeometryBasics.Rect — Type

const Rect{N,T} = HyperRectangle{N,T}

A rectangle in N dimensions, formally the cartesian product of intervals. See also HyperRectangle. Its aliases are

T(eltype)	Float64	Float32	Int

N(dim)	Rect{N,T}	Rectd{N}	Rectf{N}	Recti{N}

2	Rect2{T}	Rect2d	Rect2f	Rect2i

3	Rect3{T}	Rect3d	Rect3f	Rect3i

There is an additional unexported alias RectT that simply reverses the order of type parameters: RectT{T,N} == Rect{N,T}.

GeometryBasics.Rect — Method

Rect(vals::Number...)

Rect(vals::Number...)

Rect constructor for individually specified intervals. e.g. Rect(0,0,1,2) has origin == Vec(0,0) and width == Vec(1,2)

GeometryBasics.Rect — Method

Construct a HyperRectangle enclosing all points.

GeometryBasics.Rect2 — Type

const Rect{N,T} = HyperRectangle{N,T}

A rectangle in N dimensions, formally the cartesian product of intervals. See also HyperRectangle. Its aliases are

T(eltype)	Float64	Float32	Int

N(dim)	Rect{N,T}	Rectd{N}	Rectf{N}	Recti{N}

2	Rect2{T}	Rect2d	Rect2f	Rect2i

3	Rect3{T}	Rect3d	Rect3f	Rect3i

There is an additional unexported alias RectT that simply reverses the order of type parameters: RectT{T,N} == Rect{N,T}.

GeometryBasics.Rect2d — Type

const Rect{N,T} = HyperRectangle{N,T}

A rectangle in N dimensions, formally the cartesian product of intervals. See also HyperRectangle. Its aliases are

T(eltype)	Float64	Float32	Int

N(dim)	Rect{N,T}	Rectd{N}	Rectf{N}	Recti{N}

2	Rect2{T}	Rect2d	Rect2f	Rect2i

3	Rect3{T}	Rect3d	Rect3f	Rect3i

There is an additional unexported alias RectT that simply reverses the order of type parameters: RectT{T,N} == Rect{N,T}.

GeometryBasics.Rect2f — Type

const Rect{N,T} = HyperRectangle{N,T}

A rectangle in N dimensions, formally the cartesian product of intervals. See also HyperRectangle. Its aliases are

T(eltype)	Float64	Float32	Int

N(dim)	Rect{N,T}	Rectd{N}	Rectf{N}	Recti{N}

2	Rect2{T}	Rect2d	Rect2f	Rect2i

3	Rect3{T}	Rect3d	Rect3f	Rect3i

There is an additional unexported alias RectT that simply reverses the order of type parameters: RectT{T,N} == Rect{N,T}.

GeometryBasics.Rect2i — Type

const Rect{N,T} = HyperRectangle{N,T}

A rectangle in N dimensions, formally the cartesian product of intervals. See also HyperRectangle. Its aliases are

T(eltype)	Float64	Float32	Int

N(dim)	Rect{N,T}	Rectd{N}	Rectf{N}	Recti{N}

2	Rect2{T}	Rect2d	Rect2f	Rect2i

3	Rect3{T}	Rect3d	Rect3f	Rect3i

There is an additional unexported alias RectT that simply reverses the order of type parameters: RectT{T,N} == Rect{N,T}.

GeometryBasics.Rect3 — Type

const Rect{N,T} = HyperRectangle{N,T}

A rectangle in N dimensions, formally the cartesian product of intervals. See also HyperRectangle. Its aliases are

T(eltype)	Float64	Float32	Int

N(dim)	Rect{N,T}	Rectd{N}	Rectf{N}	Recti{N}

2	Rect2{T}	Rect2d	Rect2f	Rect2i

3	Rect3{T}	Rect3d	Rect3f	Rect3i

There is an additional unexported alias RectT that simply reverses the order of type parameters: RectT{T,N} == Rect{N,T}.

GeometryBasics.Rect3d — Type

const Rect{N,T} = HyperRectangle{N,T}

A rectangle in N dimensions, formally the cartesian product of intervals. See also HyperRectangle. Its aliases are

T(eltype)	Float64	Float32	Int

N(dim)	Rect{N,T}	Rectd{N}	Rectf{N}	Recti{N}

2	Rect2{T}	Rect2d	Rect2f	Rect2i

3	Rect3{T}	Rect3d	Rect3f	Rect3i

There is an additional unexported alias RectT that simply reverses the order of type parameters: RectT{T,N} == Rect{N,T}.

GeometryBasics.Rect3f — Type

const Rect{N,T} = HyperRectangle{N,T}

A rectangle in N dimensions, formally the cartesian product of intervals. See also HyperRectangle. Its aliases are

T(eltype)	Float64	Float32	Int

N(dim)	Rect{N,T}	Rectd{N}	Rectf{N}	Recti{N}

2	Rect2{T}	Rect2d	Rect2f	Rect2i

3	Rect3{T}	Rect3d	Rect3f	Rect3i

There is an additional unexported alias RectT that simply reverses the order of type parameters: RectT{T,N} == Rect{N,T}.

GeometryBasics.Rect3i — Type

const Rect{N,T} = HyperRectangle{N,T}

A rectangle in N dimensions, formally the cartesian product of intervals. See also HyperRectangle. Its aliases are

T(eltype)	Float64	Float32	Int

N(dim)	Rect{N,T}	Rectd{N}	Rectf{N}	Recti{N}

2	Rect2{T}	Rect2d	Rect2f	Rect2i

3	Rect3{T}	Rect3d	Rect3f	Rect3i

There is an additional unexported alias RectT that simply reverses the order of type parameters: RectT{T,N} == Rect{N,T}.

GeometryBasics.Rectd — Type

const Rect{N,T} = HyperRectangle{N,T}

A rectangle in N dimensions, formally the cartesian product of intervals. See also HyperRectangle. Its aliases are

T(eltype)	Float64	Float32	Int

N(dim)	Rect{N,T}	Rectd{N}	Rectf{N}	Recti{N}

2	Rect2{T}	Rect2d	Rect2f	Rect2i

3	Rect3{T}	Rect3d	Rect3f	Rect3i

There is an additional unexported alias RectT that simply reverses the order of type parameters: RectT{T,N} == Rect{N,T}.

GeometryBasics.Rectf — Type

const Rect{N,T} = HyperRectangle{N,T}

A rectangle in N dimensions, formally the cartesian product of intervals. See also HyperRectangle. Its aliases are

T(eltype)	Float64	Float32	Int

N(dim)	Rect{N,T}	Rectd{N}	Rectf{N}	Recti{N}

2	Rect2{T}	Rect2d	Rect2f	Rect2i

3	Rect3{T}	Rect3d	Rect3f	Rect3i

There is an additional unexported alias RectT that simply reverses the order of type parameters: RectT{T,N} == Rect{N,T}.

GeometryBasics.Recti — Type

const Rect{N,T} = HyperRectangle{N,T}

A rectangle in N dimensions, formally the cartesian product of intervals. See also HyperRectangle. Its aliases are

T(eltype)	Float64	Float32	Int

N(dim)	Rect{N,T}	Rectd{N}	Rectf{N}	Recti{N}

2	Rect2{T}	Rect2d	Rect2f	Rect2i

3	Rect3{T}	Rect3d	Rect3f	Rect3i

There is an additional unexported alias RectT that simply reverses the order of type parameters: RectT{T,N} == Rect{N,T}.

GeometryBasics.Simplex — Type

A Simplex is a generalization of an N-dimensional tetrahedra and can be thought of as a minimal convex set containing the specified points.

A 0-simplex is a point.

A 1-simplex is a line segment.

A 2-simplex is a triangle.

A 3-simplex is a tetrahedron.

Note that this datatype is offset by one compared to the traditional mathematical terminology. So a one-simplex is represented as Simplex{2,T}. This is for a simpler implementation.

It applies to infinite dimensions. The structure of this type is designed to allow embedding in higher-order spaces by parameterizing on T.

GeometryBasics.Sphere — Type

Sphere{T}

An alias for a HyperSphere of dimension 3. (i.e. HyperSphere{3, T})

GeometryBasics.Tesselation — Type

Tesselation(primitive, nvertices)

For abstract geometries, when we generate a mesh from them, we need to decide how fine grained we want to mesh them. To transport this information to the various decompose methods, you can wrap it in the Tesselation object e.g. like this:

sphere = Sphere(Point3f(0), 1)
m1 = mesh(sphere) # uses a default value for tesselation
m2 = mesh(Tesselation(sphere, 64)) # uses 64 for tesselation
length(coordinates(m1)) != length(coordinates(m2))

For grid based tesselation, you can also use a tuple: ```julia rect = Rect2(0, 0, 1, 1) Tesselation(rect, (5, 5))

GeometryBasics.TupleView — Type

TupleView{T, N, Skip, A}

TupleView, groups elements of an array as tuples. N is the dimension of the tuple, M tells how many elements to skip to the next tuple. By default TupleView{N} defaults to skip N items.

a few examples:

x = [1, 2, 3, 4, 5, 6]
TupleView{2, 1}(x):
> [(1, 2), (2, 3), (3, 4), (4, 5), (5, 6)]

TupleView{2}(x):
> [(1, 2), (3, 4), (5, 6)]

TupleView{2, 3}(x):
> [(1, 2), (4, 5)]

TupleView{3, 1}(x):
> [(1, 2, 3), (2, 3, 4), (3, 4, 5), (4, 5, 6)]

TupleView can be used together with reinterpret:

x = [1, 2, 3, 4]
y = reinterpret(Point{2, Int}, TupleView{2, 1}(x))
> [Point(1, 2), Point(2, 3), Point(3, 4)]

GeometryBasics.area — Method

area(contour::AbstractVector{Point}})

Calculate the area of a polygon.

For 2D points, the oriented area is returned (negative when the points are oriented clockwise).

GeometryBasics.area — Method

area(vertices::AbstractVector{Point{3}}, faces::AbstractVector{TriangleFace})

Calculate the area of all triangles.

GeometryBasics.area — Method

area(vertices::AbstractVector{Point{3}}, face::TriangleFace)

Calculate the area of one triangle.

GeometryBasics.connect — Method

connect(points::AbstractVector{<: Point}, P::Type{<: Polytope{N}}, skip::Int = N)

Creates a view that connects a number of points to a Polytope P. Between each polytope, skip elements are skipped untill the next starts. Example: ```julia x = connect(Point[(1, 2), (3, 4), (5, 6), (7, 8)], Line, 2) x == [Line(Point(1, 2), Point(3, 4)), Line(Point(5, 6), Point(7, 8))]

GeometryBasics.convert_simplex — Method

convert_simplex(::Type{Face{2}}, f::Face{N})

Extract all line segments in a Face.

GeometryBasics.convert_simplex — Method

convert_simplex(::Type{Face{3}}, f::Face{N})

Triangulate an N-Face into a tuple of triangular faces.

GeometryBasics.coordinates — Method

coordinates(geometry)

Returns the edges/vertices/coordinates of a geometry. Is allowed to return lazy iterators! Use decompose(ConcretePointType, geometry) to get Vector{ConcretePointType} with ConcretePointType to be something like Point{3, Float32}.

GeometryBasics.decompose — Method

decompose(facetype, contour::AbstractArray{<:Point})

Triangulate a Polygon without hole.

Returns a Vector{facetype} defining indexes into contour.

GeometryBasics.faces — Method

faces(geometry)

Returns the face connections of a geometry. Is allowed to return lazy iterators! Use decompose(ConcreteFaceType, geometry) to get Vector{ConcreteFaceType} with ConcreteFaceType to be something like TriangleFace{Int}.

GeometryBasics.normals — Method

normals{VT,FD,FT,FO}(vertices::Vector{Point{3, VT}},
                    faces::Vector{Face{FD,FT,FO}},
                    NT = Normal{3, VT})

Compute all vertex normals.

GeometryBasics.self_intersections — Method

self_intersections(points::AbstractVector{<:Point})

Finds all self intersections of polygon points

GeometryBasics.split_intersections — Method

split_intersections(points::AbstractVector{<: Point})

Splits polygon points into it's self intersecting parts. Only 1 intersection is handled right now.

GeometryBasics.volume — Method

volume(mesh)

Calculate the signed volume of all tetrahedra. Be sure the orientation of your surface is right.

GeometryBasics.volume — Method

volume(triangle)

Calculate the signed volume of one tetrahedron. Be sure the orientation of your surface is right.

Private

GeometryBasics.Ngon — Type

Fixed Size Polygon, e.g.

N 1-2 : Illegal!

N = 3 : Triangle

N = 4 : Quadrilateral (or Quad, Or tetragon)

N = 5 : Pentagon

...

GeometryBasics.RectT — Type

const Rect{N,T} = HyperRectangle{N,T}

A rectangle in N dimensions, formally the cartesian product of intervals. See also HyperRectangle. Its aliases are

T(eltype)	Float64	Float32	Int

N(dim)	Rect{N,T}	Rectd{N}	Rectf{N}	Recti{N}

2	Rect2{T}	Rect2d	Rect2f	Rect2i

3	Rect3{T}	Rect3d	Rect3f	Rect3i

There is an additional unexported alias RectT that simply reverses the order of type parameters: RectT{T,N} == Rect{N,T}.

Base.:* — Method

*(m::Mat, h::Rect)

Transform a Rect using a matrix. Maintains axis-align properties so a significantly larger Rect may be generated.

Base.in — Method

in(b1::Rect, b2::Rect)

Check if Rect b1 is contained in b2. This does not use strict inequality, so Rects may share faces and this will still return true.

Base.in — Method

in(pt::VecTypes, b1::Rect{N, T})

Check if a point is contained in a Rect. This will return true if the point is on a face of the Rect.

Base.in — Method

in(point, triangle)

Determine if a point is inside of a triangle.

Base.intersect — Method

intersect(h1::Rect, h2::Rect)

Perform a intersection between two Rects.

Base.isempty — Method

isempty(h::Rect)

Return true if any of the widths of h are negative.

Base.union — Method

union(r1::Rect{N}, r2::Rect{N})

Returns a new Rect{N} which contains both r1 and r2.

GeometryBasics.mesh — Method

mesh(primitive::GeometryPrimitive;
     pointtype=Point, facetype=GLTriangle,
     uv=nothing, normaltype=nothing)

Creates a mesh from primitive.

Uses the element types from the keyword arguments to create the attributes. The attributes that have their type set to nothing are not added to the mesh. Note, that this can be an Int or Tuple{Int, Int}`, when the primitive is grid based. It also only losely correlates to the number of vertices, depending on the algorithm used. #TODO: find a better number here!

GeometryBasics.mesh — Method

mesh(polygon::AbstractVector{P}; pointtype=P, facetype=GLTriangleFace,
     normaltype=nothing)

Create a mesh from a polygon given as a vector of points, using triangulation.

GeometryBasics.orthogonal_vector — Method

The unnormalized normal of three vertices.

GeometryBasics.split — Method

split(rectangle, axis, value)

Splits an Rect into two along an axis at a given location.